+++ azalea-client/src/plugins/mining.rs
@@ -55,7 +55,6 @@ impl Plugin for MiningPlugin {
                     .in_set(MiningSet)
                     .after(InventorySet)
                     .after(MoveEventsSet)
-                    .before(azalea_entity::update_bounding_box)
                     .after(azalea_entity::update_fluid_on_eyes)
                     .after(crate::interact::pick::update_hit_result_component)
                     .after(crate::attack::handle_attack_event)