+++ azalea-entity/src/data.rs
@@ -152,9 +152,19 @@ pub enum Pose {
     Sleeping,
     Swimming,
     SpinAttack,
-    Sneaking,
+    Crouching,
     LongJumping,
     Dying,
+    Croaking,
+    UsingTongue,
+    Sitting,
+    Roaring,
+    Sniffing,
+    Emerging,
+    Digging,
+    Sliding,
+    Shooting,
+    Inhaling,
 }
 
 #[derive(Debug, Clone, AzBuf)]